test: extract helper fixtures to the conftest.py
